Attack Tools special settings (no language data)
This implementation allows modifying each single attack, but for one I feel like it is too bothersome to be used and it comes with a few problems:
Attacks are not (necessarily) labeled, they are simply a list. If the attack is changed by the original author, for example an attack is removed from the list, changes would be applied to the next attack instead, potentially causing issues or at least weird outcomes.
I'm also unhappy and unsure about how the list is saved when changes are made and how it is put into the save despite no changes being made.

Of course, this way offers more freedom in what can be modified.
With further development it could allow for changing the type of damage and ExtraDamages.

Overall I feel that putting in two simple settings for modifying attack damage and speed might be the easier way to handle it though.