Egg Layer setting allows enabling or disabling egg laying for all animals -- see issues!
Issues:
- ThingDefs for eggs are added programatically; once (unfert.) egg laying is activated for an animal, a def for the egg will be added if it does not already exist.
This ThingDef will be removed if the egg laying is disabled. Upon loading a game where eggs using a now-removed ThingDef already exist, the game will throw errors. Unsure how problematic that is, I was going to test it with meals using eggs that then get removed but that lead me to a much worse problem...
- Programatically generated egg-ThingDefs are not added to stockpiles or recipes (etc.). I'm probably missing something that makes this happen or such...
I saw two ways for adding the Defs, either they only get added when needed, as in, when an animal is set to use the egg-def, or they are generated at the start for every animal, which causes lots of unused def clutter (two defs per animal, fertilized and unfertilized). While the second option would allow for generating the egg-defs at the start of the game and probably fix the issue, it still leaves me questioning whether there aren't other issues I don't see.

So I decided to scrap this idea for now and only allow for editing existing eggs, as sad as that is.
No egg-laying Alpacas for now.

This also foreshadows limitations for wool and milk production settings...